### 2017-11-22 #159 @chrisvwn
Nightlights Data Extraction in R
The Rnightlights package extracts raster and zonal statistics from satellite nightlight rasters downloaded from the United States National Oceanic and Atmospheric Administration (<http://www.noaa.gov>) free data repositories. Statistics are calculated at the country and sub-country levels.

```
Package: Rnightlights
Version: 0.1.3
Title: Satellite Nightlight Data Extraction
Description: The Rnightlights package extracts raster and zonal statistics
    from satellite nightlight rasters downloaded from the United States
    National Oceanic and Atmospheric Administration (<http://www.noaa.gov>)
    free data repositories. Both the DMSP-OLS annual and SNPP-VIIRS monthly
    nightlight raster data are supported. Satellite nightlight raster tiles are
    downloaded and cropped to the country boundaries using shapefiles from the GADM
    database of Global Administrative Areas (<http://gadm.org>). Zonal statistics
    are then calculated at the lowest administrative boundary for the selected
    country and cached locally for future retrieval. Finally, a simple data
    explorer/browser is included that allows one to visualize the cached data e.g.
    graphing, mapping and clustering regional data.
Authors@R: c(person(given="Christopher", family="Njuguna", role=c("aut", "cre", "cph"), email="chris.njuguna@gmail.com"),
            person(given="Henrik", family="Bengtsson", role="ctb", email="henrikb@braju.com", comment="data setup, https://cran.r-project.org/package=R.cache"),
            person(given="Black", family="Guru", role="ctb", comment="gdal zonal, http://www.guru-gis.net/efficient-zonal-statistics-using-r-and-gdal/"),
            person(given="Jeff", family="Chen", role="ctb", comment="masq function, https://commercedataservice.github.io/tutorial_viirs_part1/"),
            person(given="Star", family="Ying", role="ctb", comment="masq function, https://commercedataservice.github.io/tutorial_viirs_part1/"),
            person(given="Chris", family="Elvidge", role="ctb", comment="information + masq function"))
    cleangeo,
    compiler,
    doParallel,
    foreach,
    gdalUtils,
    raster,
    reshape2,
    rgdal,
    rgeos,
    R.utils,
    rvest,
    rworldmap,
    settings,
    sp,
    xml2
    dendextend,
    ggdendro,
    leaflet,
    plotly,
    RColorBrewer,
    reshape,
    shiny,
    shinydashboard
SystemRequirements:
    gdal
    wget
    curl
    aria2

```

http://github.com/chrisvwn/Rnightlights

[e.g., "data extraction, because the package parses a scientific data file format"]
data retrieval, because the package downloads the appropriate satellite nightlight raster tiles given time period and country ISO3 code
data extraction, because the package clips satellite nightlight imagery to country boundaries and performs aggregation functions on the rasters
geospatial data, because the package deals with satellite nightlight geospatial raster tiles

Researchers and anyone who would like to perform analyses of satellite nightlights imagery.

There are no R packages that accomplish the same thing that I am aware of.
