+++ README.md
@@ -25,9 +25,10 @@ devtools::install_github("chrisvwn/Rnightlights")
 An example to process VIIRS monthly nightlights for Kenya for the year 2014 from the
 [Ishara Data blog post](http://isharadata.blogspot.co.ke/2017/09/rnightlights-satellite-nightlight-data.html)
 
-Note 1: Change the "ctry" variable to run the example on another country
-Note 2: This will not work for countries without admin levels below the country level e.g. ATA (Antarctica)
-Note 3: This calculates total radiances per region and so may be biased by area. Normalize by area to see 
+####Notes
+1. Change the "ctry" variable to run the example on another country
+2. This will not work for countries without admin levels below the country level e.g. ATA (Antarctica)
+3. This calculates total radiances per region and so may be biased by area. Normalize by area to see 
         regions with higher radiances per unit area e.g. to estimate areas with higher economic activity
 
 ```{r}
