### 2017-11-22 cbc0375
Updated code and documentation to reflect change of the earliest VIIRS data available to "201401" instead of the previous "201204"
