### 2017-11-22 4051e5d
Removed reference to VIIRS periods before 201401
