### 2017-11-22 c03ca68
Updated version and documentation
