### 2017-11-24 158cd87
Updating example to add "ignoreMissing=FALSE" to the getCtryNlData function to work on new installs
