### 2017-11-24 #159 @sckott
@chrisvwn Submission is a good fit for rOpenSci. We'll assign an editor asap
