### 2017-11-29 #159 @maelle
@chrisvwn before starting the review process your package must:

* have a test suite 😉  Do you need help with this? I can recommend reading http://r-pkgs.had.co.nz/tests.html and then ask me any question, for that open an issue in your repo (to keep this thread clean) or create a thread in [our forum](https://discuss.ropensci.org/). Don't worry, we'll get you there!

* have a low case name as advised in our guidelines (since your package isn't on CRAN yet and I imagine doesn't have a large userbase yet?): it's easier for users to remember e.g. "rnightlights" and for the sake of simplicity you could even call it "nightlights". I changed the name of my own "Ropenaq" package to "ropenaq" to comply and I can say it wasn't very difficult.

I'll put this submission on hold because of this.

I also have 3 remarks/questions:

* Any reason why you use `reshape` and not `tidyr` or `reshape2`?

* Add README.md, the Travis config file etc to .Rbuildignore

* I'd rephrase
```
Enter 0 to use a temporary directory for this session only.

1: Create data path under home directory 'C:/Users/Maelle/Documents
2: Choose a different directory as the data path
```

to

```
0: Use a temporary directory for this session only.
1: Create data path under home directory 'C:/Users/Maelle/Documents
2: Choose a different directory as the data path
```
