### 2017-11-30 #159 @maelle
Oh sorry,  if it is on CRAN don't change the name ☺
