### 2017-12-06 e502c4d
Return VIIRS start period to 201401
