### 2017-12-07 2ad420c
Export stats functions in doParallel::foreach to avoid function not found in Windows where functions are defined in the parent environment
