### 2017-12-07 a50a7b1
switched raster parallel statistics calculation from doParallel to doSNOW to enable progressbar output at the console
