+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(Rnightlights)
+
+test_check("Rnightlights")
