### 2017-12-09 453c843
+Updated description +Removed contributors who should actually be references
