### 2017-12-09 91bf497
Included missing dependencies
