+++ DESCRIPTION
@@ -19,7 +19,7 @@ Imports:
     cleangeo,
     compiler,
     data.table,
-    doParallel,
+    doSNOW,
     dplyr,
     foreach,
     gdalUtils,
@@ -34,6 +34,7 @@ Imports:
     rvest,
     rworldmap,
     settings,
+    snow,
     sp,
     stats,
     stringr,
