### 2017-12-21 d99c2d8
Changed wildcard for finding the average dnb tif in the tgz to reflect naming change. From avg_rade9.tif is now avg_rade9h.tif
