+++ man/processNlData.Rd
@@ -58,7 +58,7 @@ This is the main entry-point to the package and the main user-facing function. H
 
     nlPeriods <- getAllNlYears("VIIRS") #get a list of all nightlight periods to present-day
 
-    nlPeriods <- nlPeriods[grep("^2012", nlPeriods)] #filter only periods in 2012
+    nlPeriods <- nlPeriods[grep("^2014", nlPeriods)] #filter only periods in 2014
 
     processNlData(nlPeriods=nlPeriods)
 
