### 2017-12-23 24c8a40
Changed download path to the data dir instead of root folder of working drive
