### 2017-12-24 d2da5c7
First successful try using ffdf and ffdfply to calculate stats. Slower than the partial calculation but accurate for all stats by passing zonal stats using disk to allow for large rasters
