### 2017-12-25 8232a8b
Implemented myZonal using ffdfs to enable storage of large raster and zonal files to disk. Slower than the partial calculations which only works for functions like sum which can take streaming data. But faster than parallel raster calculations on polygons.
