### 2017-12-25 d655b12
Removed old myZonals and related function implementations that are no longer in use
