+++ R/stats.R
@@ -39,120 +39,6 @@ validStat <- function(stat)
     return(TRUE)
 }
 
-raster2FFDF <- function(rast, zone, ...)
-{
-  message("Reading in raster data")
-  
-  file.remove("rast.ff", "zone.ff")
-  
-  nc <- ncol(rast)
-  
-  tr <- raster::blockSize(rast)
-  
-  pb <- utils::txtProgressBar(min=1, max=tr$n, style=3)
-  
-  for (i in 1:tr$n)
-  {
-    start <- ((tr$row[i]-1) * nc) + 1
-    
-    end <- start + (tr$nrows[i] * nc) - 1
-    
-    rastVals <- raster::getValues(rast, row=tr$row[i], nrows=tr$nrows[i])
-    zoneVals <- raster::getValues(zone, row=tr$row[i], nrows=tr$nrows[i])
-    
-    rastVals[rastVals < 0] <- NA
-    
-    idxZone0 <- which(zoneVals == 0)
-    
-    #zoneVals <- zoneVals[-idxZone0]
-    #rastVals <- rastVals[-idxZone0]
-    
-    if (i == 1)
-    {
-      vals <- ff::ff(initdata = rastVals, filename = "rast.ff", overwrite = T)
-      zones <- ff::ff(initdata = zoneVals, filename = "zone.ff", overwrite = T)
-    } 
-    else 
-    {
-      vals <- ffbase::ffappend(vals, rastVals, adjustvmode = T)
-      zones <- ffbase::ffappend(zones, zoneVals, adjustvmode = T)
-    }
-    
-    utils::setTxtProgressBar(pb, i)
-  }
-  
-  close(pb)
-  
-  rDT <- ff::ffdf(zones, vals)
-  
-  rDT
-}
-
-raster2FFDF1 <- function(rast, filename="")
-{
-  message("Reading in raster data")
-  
-  #file.remove("rast.ff", "zone.ff")
-  
-  nc <- ncol(rast)
-  
-  tr <- raster::blockSize(rast)
-  
-  pb <- utils::txtProgressBar(min=1, max=tr$n, style=3)
-  
-  for (i in 1:tr$n)
-  {
-    start <- ((tr$row[i]-1) * nc) + 1
-    
-    end <- start + (tr$nrows[i] * nc) - 1
-    
-    rastVals <- raster::getValues(rast, row=tr$row[i], nrows=tr$nrows[i])
-    #zoneVals <- raster::getValues(zone, row=tr$row[i], nrows=tr$nrows[i])
-    
-    #rastVals[rastVals < 0] <- NA
-    
-    #idxZone0 <- which(zoneVals == 0)
-    
-    #zoneVals <- zoneVals[-idxZone0]
-    #rastVals <- rastVals[-idxZone0]
-    
-    if (i == 1)
-    {
-      vals <- ff::ff(initdata = rastVals, filename = filename, overwrite = T)
-      #zones <- ff::ff(initdata = zoneVals, filename = "zone.ff", overwrite = T)
-    } 
-    else 
-    {
-      vals <- ffbase::ffappend(vals, rastVals, adjustvmode = T)
-      #zones <- ffbase::ffappend(zones, zoneVals, adjustvmode = T)
-    }
-    
-    utils::setTxtProgressBar(pb, i)
-  }
-  
-  close(pb)
-  
-  #rDT <- ff::ffdf(zones, vals)
-  
-  vals
-}
-
-raster2BigMat <- function(rast, filename="") {
-  b <- bigmemory::big.matrix(raster::ncell(rast), raster::nlayers(rast), backingfile=filename, descriptorfile = paste0(filename, ".desc") )
-  
[…]
