### 2017-12-25 6f07034
Added full package references of function examples
