### 2017-12-26 c5a03f3
Modified gdal data extraction to use ffdf method
