+++ R/ctrynldata.R
@@ -10,7 +10,7 @@
 #' @return dataframe with the country admin level data
 #'
 #' @examples
-#' \dontrun{initCtryNlData <- createCtryNlDataDF("KEN")}
+#' \dontrun{initCtryNlData <- Rnightlights::createCtryNlDataDF("KEN")}
 #'  #returns a data frame
 #'
 createCtryNlDataDF <- function(ctryCode)
@@ -103,14 +103,15 @@ createCtryNlDataDF <- function(ctryCode)
 #' 
 #' @param nlType the type of nightlight data i.e. "OLS" or "VIIRS"
 #'
-#' @return Character full path to the cropped VIIRS country raster 
-#'     for a country and a given year and month
+#' @return the updated dataframe
 #'
 #' @examples
 #' 
-#' \dontrun{ctryNlDataDF <- insertNlDataCol(ctryNlDataDF, dataCol, "sum", "201409", "VIIRS")}
+#' \dontrun{ctryNlDataDF <- Rnightlights::insertNlDataCol(ctryNlDataDF, 
+#'     dataCol, "sum", "201409", "VIIRS")}
 #' 
-#' \dontrun{ctryNlDataDF <- insertNlDataCol(ctryNlDataDF, dataCol, "mean", "2012", "OLS")}
+#' \dontrun{ctryNlDataDF <- Rnightlights::insertNlDataCol(ctryNlDataDF, 
+#'     dataCol, "mean", "2012", "OLS")}
 #'
 insertNlDataCol <- function (ctryNlDataDF, dataCol, statType, nlPeriod, nlType)
 {
@@ -155,6 +156,68 @@ insertNlDataCol <- function (ctryNlDataDF, dataCol, statType, nlPeriod, nlType)
   ctryNlDataDF <- ctryNlDataDF[ , newNlDataColNames]
   
   return(ctryNlDataDF)
+}
+
+######################## deleteNlDataCol ###################################
+
+#' Delete an aggregate nightlight data column in a country nightlights dataframe
+#'
+#' Delete an aggregate nightlight data column in a country nightlights dataframe. The number
+#'     of elements in the vector MUST match the number of rows in the country dataframe.
+#'
+#' @param ctryCode country with the  data column to remove
+#' 
+#' @param nlType  the type of nightlight data i.e. "OLS" or "VIIRS"
+#' 
+#' @param nlPeriod the nlPeriod that the dataCol belongs to
+#' 
+#' @param statType the stat which produced the dataCol vector
+#'
+#' @examples
+#' 
+#' \dontrun{ctryNlDataDF <- Rnightlights::deleteNlDataCol(ctryNlDataDF, 
+#'      "VIIRS", "201409", "sum")}
+#' 
+#' \dontrun{Rnightlights::deleteNlDataCol(ctryNlDataDF, 
+#'     "OLS", "2012", "mean")}
+#'
+#' @export
+deleteNlDataCol <- function (ctryCode,nlType, nlPeriod, statType)
+{
+  if(missing(ctryCode))
+    stop("Missing required parameter ctryCode")
+
+  if(missing(nlType))
+    warning("Missing required parameter nlType")
+  
+  if(missing(nlPeriod))
+    stop("Missing required parameter nlPeriod")
+  
+  if(missing(statType))
+    stop("Missing required parameter statType")
+  
+  ctryNlDataDF <- getCtryNlData(ctryCode = ctryCode, 
+                                nlPeriods = getAllNlPeriods(nlType), 
+                                nlType = nlType,
+                                ignoreMissing = TRUE)
+  
+  #read in all column names in the dataframe
+  cols <- names(ctryNlDataDF)
+  
+  colName <- paste0("NL", nlType, toupper(statType), collapse = "_")
+  
[…]
