+++ man/createCtryNlDataDF.Rd
@@ -17,7 +17,7 @@ Initiates the country nightlight dataframe with the country data read from the p
     This includes admin levels, level names and area
 }
 \examples{
-\dontrun{initCtryNlData <- createCtryNlDataDF("KEN")}
+\dontrun{initCtryNlData <- Rnightlights::createCtryNlDataDF("KEN")}
  #returns a data frame
 
 }

[…]
