### 2017-12-28 ea8de45
Changed order of params to match other functions ctryCode, nlType, nlPeriods, stats
