### 2017-12-29 bcea294
+Changing order of params in getCtryNlData +Renaming stats to nlStats to match other nlxxx names and avoid ambiguity with stats package +Renamed gdal_cachemax pkgOption to camelcase gdalCacheMax +Changed hardcoded gdalwarp option NUM_THREADS to use pkgOption numCores
