+++ R/nightlights.R
@@ -89,8 +89,8 @@
 #' @param extractMethod ("rast" or "gdal") Whether to use rasterize or gdal-based functions 
 #'     to crop and mask the country rasters
 #' 
-#' @param fnStats the statistics to calculate. If not provided will calculate the stats specified 
-#'     in \code{pkgOptions("stats")}
+#' @param nlStats the statistics to calculate. If not provided will calculate the stats specified 
+#'     in \code{pkgOptions("nlStats")}
 #'
 #' @return None
 #'
@@ -100,7 +100,7 @@
 #' #for both cropMask and extraction for KEN
 #' \dontrun{processNLCountry("KEN", "201412", "VIIRS", "gdal", "gdal", "sum")}
 #'
-processNLCountry <- function(ctryCode, nlPeriod, nlType, cropMaskMethod=pkgOptions("cropMaskMethod"), extractMethod=pkgOptions("extractMethod"), fnStats=pkgOptions("stats"))
+processNLCountry <- function(ctryCode, nlPeriod, nlType, cropMaskMethod=pkgOptions("cropMaskMethod"), extractMethod=pkgOptions("extractMethod"), nlStats=pkgOptions("nlStats"))
 {
   if(missing(ctryCode))
     stop("Missing required parameter ctryCode")
@@ -130,7 +130,7 @@ processNLCountry <- function(ctryCode, nlPeriod, nlType, cropMaskMethod=pkgOptio
   {
     message("Data file found: ", getCtryNlDataFnamePath(ctryCode))
     
-    if(all(sapply(fnStats, function(stat) existsCtryNlData(ctryCode, nlPeriod, stat, nlType))))
+    if(all(sapply(nlStats, function(nlStat) existsCtryNlData(ctryCode, nlPeriod, nlStat, nlType))))
     {
       message("All stats exist for ", ctryCode, " ", nlPeriod, ". Skipping")
 
@@ -230,18 +230,18 @@ processNLCountry <- function(ctryCode, nlPeriod, nlType, cropMaskMethod=pkgOptio
       #GDALWARP
       rstTmp <- file.path(getNlDir("dirNlTiles"), paste0(basename(tempfile()), ".tif"))
       
-      message("Writing merged raster to disk for gdalwarp masking")
+      message("Writing merged raster to disk for gdalwarp masking", base::date())
       
       raster::writeRaster(ctryRastCropped, rstTmp)
       
-      output_file_vrt <- file.path(getNlDataPath(), paste0(ctryCode, "_", nlType, "_", nlPeriod, ".vrt"))
+      output_file_vrt <- file.path(tempdir(), paste0(ctryCode, "_", nlType, "_", nlPeriod, ".vrt"))
       
       if (file.exists(output_file_vrt))
         file.remove(output_file_vrt)
       
       message("gdalwarp masking",base::date())
       
-      gdalUtils::gdalwarp(srcfile=rstTmp, dstfile=output_file_vrt, s_srs=wgs84, t_srs=wgs84, cutline=getPolyFnamePath(ctryCode), cl= getCtryShpLyrName(ctryCode,0), multi=TRUE, wm=2048, wo="NUM_THREADS=ALL_CPUS")
+      gdalUtils::gdalwarp(srcfile=rstTmp, dstfile=output_file_vrt, s_srs=wgs84, t_srs=wgs84, cutline=getPolyFnamePath(ctryCode), cl= getCtryShpLyrName(ctryCode,0), multi=TRUE, wm=pkgOptions("gdalCacheMax"), wo=paste0("NUM_THREADS=", pkgOptions("numCores")))
       
       message("gdal_translate ", base::date())
[…]

[…]
