### 2017-12-29 d0e68c6
Re-enabled removal of global vars on package detach
