+++ R/zzz.R
@@ -25,8 +25,8 @@
 
 .onDetach <- function(libname)
 {
-  #remove any global vars we created in .onLoad
-  #suppressWarnings(rm(map, shpTopLyrName, wgs84, nlTiles, tilesSpPolysDFs))
+  #remove any global vars we created in .onAttach
+  suppressWarnings(rm(map, shpTopLyrName, wgs84, nlTiles, tilesSpPolysDFs))
   
   #cleanup by removing any global vars created etc
   nlCleanup();
