### 2017-12-29 77fda47
+Various miscellaneous
