### 2018-01-02 c3ef75d
Dynamic nlType and nlStats in the exploreData UI
