+++ inst/application/server.R
@@ -75,7 +75,7 @@ if (!requireNamespace("rgdal", quietly = TRUE))
 
 if (!requireNamespace("Rnightlights", quietly = TRUE))
 {
-  missingPkgs <- c(missingPkgs, "rstudio/DT")
+  missingPkgs <- c(missingPkgs, "Rnightlights")
 }
 
 if(!is.null(missingPkgs))
@@ -130,30 +130,102 @@ shiny::shinyServer(function(input, output, session){
       
       data <- data.table::fread(Rnightlights::getCtryNlDataFnamePath(countries), colClasses = colClasses, header = T)
     })
+  
+  ######################## reactive ctryNlTypes ###################################
+  
+  ctryNlTypes <- shiny::reactive({
+    #print(paste0("here: ctryNlTypes"))
     
+    countries <- input$countries
+    
+    if (length(countries) < 1)
+     return()
+    
+    #temp <- data.table::fread(Rnightlights::getCtryNlDataFnamePath(input$countries), nrows = 1, header = T)
+    
+    #temp <- ctryNlData()
+    
+    nlTypes <- NULL
+    
+    if (length(countries) == 1)
+    {
+      hdrs <- data.table::fread(Rnightlights::getCtryNlDataFnamePath(countries), nrows = 1, header = T)
+      
+      cols <- grep(pattern = "NL_", x = names(hdrs), value = T)
+      
+      nlTypes <- list(unique(sapply(cols, function(x)unlist(strsplit(x, "_"))[2])))
+    }
+    else if(length(countries) > 1) #remove subcountry admin levels
+    {
+      for (ctryCode in countries)
+      {
+        hdrs <- data.table::fread(Rnightlights::getCtryNlDataFnamePath(ctryCode), nrows = 1, header = T)
+        
+        cols <- grep(pattern = "NL_", x = names(hdrs), value = T)
+        
+        temp <- list(unique(sapply(cols, function(x)unlist(strsplit(x, "_"))[2])))
+        
+        nlTypes <- c(nlTypes, temp)
+      }
+    }
+    
+    if(is.null(nlTypes))
+      return(NULL)
+    
+    if(length(nlTypes) == 1)
+      nlTypes <- unlist(nlTypes)
+    else
+      nlTypes <- unlist(Reduce(intersect, nlTypes))
+
+    return(nlTypes)
+  })
+  
   ######################## reactive ctryDataStats ###################################
   
   ctryDataStats <- shiny::reactive({
     #print(paste0("here: ctryDataStats"))
+    #print(paste0("here: ctryNlTypes"))
     
-    # if (length(input$countries) != 1)
-    #   return()
+    countries <- input$countries
     
-    #temp <- data.table::fread(Rnightlights::getCtryNlDataFnamePath(input$countries), nrows = 1, header = T)
+    if (length(countries) < 1)
+      return()
     
-    temp <- ctryNlData()
+    nlStats <- NULL
     
-    if(is.null(temp))
+    if (length(countries) == 1)
+    {
+      hdrs <- data.table::fread(Rnightlights::getCtryNlDataFnamePath(countries), nrows = 1, header = T)
+      
+      cols <- grep(pattern = "NL_", x = names(hdrs), value = T)
+      
+      nlStats <- list(unique(gsub(".*._.*._.*._", "", cols)))
+    }
+    else if(length(countries) > 1) #remove subcountry admin levels
+    {
+      for (ctryCode in countries)
+      {
+        hdrs <- data.table::fread(Rnightlights::getCtryNlDataFnamePath(ctryCode), nrows = 1, header = T)
+        
+        cols <- grep(pattern = "NL_", x = names(hdrs), value = T)
[…]
