### 2018-01-02 01a9574
+Added some progress bars +Fixed duplication of stats processing by not reprocessing existing stats
