### 2018-01-02 188068e
Reduced gdalCacheMax from 2000 to 1024
