### 2018-01-02 827f33e
+Cleaning up comments, parameters and code
