### 2018-01-02 12425bf
Check for nlPeriod length > 1
