### 2018-01-02 f081365
Changed parameter order for processNlData
