### 2018-01-03 4cfb568
Changed the stats parameter to nlStats in the example
