### 2018-01-03 93f510e
Set global vars to NULL before deleting in .onDetach
