### 2018-01-03 f51804f
Made nlStat choices dependent on nlType
