### 2018-01-03 322feaa
Commented out web raster code
