### 2018-01-03 e221f6c
roxygenize updated man pages
