### 2018-01-03 cc0e1b5
Created a temp dir under the root data dir for use by ff and raster
