### 2018-01-04 79ae8b4
+Version 0.1.5 finalized +Updated stats and fnStats params to nlStats for consistency +Fully referenced unexported examples
