+++ tests/testthat/test-ctrycodes.R
@@ -3,7 +3,8 @@ library(Rnightlights)
 context("ctryCodes")
 
 test_that("ctryCode lookups work", {
-  expect_error(ctryCodeToName("KE"), "Only 3-letter ISO3 codes allowed", fixed=TRUE)
+  expect_warning(ctryCodeToName("KE"), "Only 3-letter ISO3 codes allowed", fixed=TRUE)
+  expect_equal(ctryCodeToName("KE"), NA)
   expect_match(ctryCodeToName("KEN"), "Kenya", fixed=TRUE)
   
   expect_that(ctryCodeToName(), is_a("data.frame"))

[…]
