### 2018-01-06 1e6a238
Removed outfile parameter from makeCluster which was printing to the console in linux
