### 2018-01-06 d9f998d
Documentation and example updates
