+++ man/dnldCtryPoly.Rd
@@ -16,7 +16,7 @@ TRUE/FALSE Success/Failure of the download
 Download a country's polygon shapefile from \url{http://gadm.org}
 }
 \examples{
-\dontrun{
+\donttest{
 Rnightlights:::dnldCtryPoly("KEN")
 }
 
