+++ man/allValid.Rd
@@ -25,11 +25,11 @@ Check if a vector/list of values given is valid as per the given
 }
 \examples{
 
-\dontrun{
+\donttest{
 Rnightlights:::allValid(c("KEZ", "UGA", "RWA", "TZA"), Rnightlights:::validCtryCode)
 }
  
-\dontrun{
+\donttest{
 Rnightlights:::allValid(c("2012", "2015"), validNlPeriod, "OLS")
 }
 
