### 2018-01-09 c9fcd58
Added tests using testthat
