### 2018-01-10 e90728b
Debug shiny app. Fixed missing DT suggestion.
