+++ DESCRIPTION
@@ -44,6 +44,7 @@ Imports:
     xml2
 Suggests:
     dendextend,
+    DT,
     ggdendro,
     ggplot2,
     leaflet,
