### 2018-01-10 40b3936
Added codecov checks and badge
