+++ .travis.yml
@@ -8,3 +8,9 @@ addons:
     packages:
       - libgdal-dev
       - libproj-dev
+      
+r_packages:
+  - covr
+
+after_success:
+  - Rscript -e 'library(covr); codecov()'
