### 2018-01-10 4b7498e
Added codecov checks and badge
