### 2018-01-11 #159 @maelle
Closing this as discussed in the package repo! When submitting your package again after it's stable please put a link to this issue. See you soon in this repository! 😉
